requirements.txt
streamlit
pandas
reportlab
sqlite3
pillow